pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "fem-lab"
version = "1.0.0"
description = "Production-quality Finite Element Method Streamlit application"
requires-python = ">=3.10"
dependencies = [
    "streamlit==1.43.0",
    "pyvista==0.44.2",
    "numpy==1.26.4",
    "matplotlib==3.9.2",
    "scikit-fem==10.0.2",
    "scipy==1.13.1",
    "pandas==2.2.3",
    "meshio==5.3.5",
    "pillow==10.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "ruff==0.6.9",
    "mypy==1.11.2",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
